+++ packages/backend/src/api/resources/CommerceSubscription.ts
@@ -1,10 +1,12 @@
-import { type CommerceMoneyAmount } from './CommercePlan';
+import { type CommerceMoneyAmount } from '@clerk/types';
+
 import { CommerceSubscriptionItem } from './CommerceSubscriptionItem';
 import type { CommerceSubscriptionJSON } from './JSON';
 
 /**
- * @experimental This is an experimental API for the Billing feature that is available under a public beta, and the API is subject to change.
- * It is advised to pin the SDK version to avoid breaking changes.
+ * The `CommerceSubscription` object is similar to the [`CommerceSubscriptionResource`](/docs/references/javascript/types/commerce-subscription-resource) object as it holds information about a subscription, as well as methods for managing it. However, the `CommerceSubscription` object is different in that it is used in the [Backend API](https://clerk.com/docs/reference/backend-api/tag/billing/get/organizations/%7Borganization_id%7D/billing/subscription){{ target: '_blank' }} and is not directly accessible from the Frontend API.
+ *
+ * @experimental This is an experimental API for the Billing feature that is available under a public beta, and the API is subject to change. It is advised to pin the SDK version to avoid breaking changes.
  */
 export class CommerceSubscription {
   constructor(
@@ -21,19 +23,19 @@ export class CommerceSubscription {
      */
     readonly payerId: string,
     /**
-     * Unix timestamp (milliseconds) of creation.
+     * Unix timestamp (milliseconds) of when the subscription was created.
      */
     readonly createdAt: number,
     /**
-     * Unix timestamp (milliseconds) of last update.
+     * Unix timestamp (milliseconds) of when the subscription was last updated.
      */
     readonly updatedAt: number,
     /**
-     * Unix timestamp (milliseconds) when the subscription became active.
+     * Unix timestamp (milliseconds) of when the subscription became active.
      */
     readonly activeAt: number | null,
     /**
-     * Unix timestamp (milliseconds) when the subscription became past due.
+     * Unix timestamp (milliseconds) of when the subscription became past due.
      */
     readonly pastDueAt: number | null,
     /**